Jxiao/bubblegum jankle refactor (#115)
Update Bubblegum contract to be (almost) identical to @austbot's upsert branch
Merged together the nonce account and tree authority
Enforced that the decompressed mint is a PDA of the asset_id and the token_program_id
Added new column to TS indexer to represent compressed state
Fixed parsing bug in redeem
Cleaned up parsing logic/code reduction clutter

+++ contracts/programs/bubblegum/src/state/mod.rs
@@ -10,7 +10,7 @@ use metaplex_adapter::MetadataArgs;
 #[account]
 #[derive(Copy)]
 pub struct Nonce {
-    pub count: u128,
+    pub count: u64,
 }
 
 #[account]
@@ -30,10 +30,9 @@ impl Voucher {
         }
     }
 }
-
 #[event]
 pub struct NewNFTEvent {
     pub version: Version,
     pub metadata: MetadataArgs,
-    pub nonce: u128,
+    pub nonce: u64,
 }
